MTVulnerability/requirements.txt
absl-py==0.11.0
adversarial-robustness-toolbox==1.5.0
alembic==1.4.2
astunparse==1.6.3
atomicwrites==1.4.0
attrs==20.3.0
brotlipy==0.7.0
cachetools==4.2.0
Cartopy==0.17.0
certifi==2020.12.5
cffi @ file:///D:/bld/cffi_1606601395031/work
chardet==3.0.4
click==7.1.2
cloudpickle==1.4.1
cma==3.0.3
colorama==0.4.4
comet-ml==3.2.11
configobj==5.0.6
cryptography @ file:///D:/bld/cryptography_1607606432554/work
cssselect==1.1.0
cycler==0.10.0
Cython==0.29.21
cytoolz==0.11.0
dask @ file:///tmp/build/80754af9/dask-core_1607706933335/work
databricks-cli==0.11.0
decorator==4.4.2
docker==4.2.1
dulwich==0.20.15
easydict==1.9
entrypoints==0.3
enum34==1.1.10
everett==1.0.3
fets==0.5.3
ffmpeg-python==0.2.0
Flask==1.1.2
flatbuffers==1.12
future==0.18.2
gast==0.3.3
GDAL==3.0.2
gitdb==4.0.5
GitPython==3.1.3
glouton==0.7.0
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
gorilla==0.3.0
GPUtil==1.4.0
grpcio==1.32.0
h5py==2.10.0
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1593328102638/work
imagecodecs-lite @ file:///D:/bld/imagecodecs-lite_1602583207014/work
imageio @ file:///tmp/build/80754af9/imageio_1594161405741/work
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==1.0.0
jsonschema==3.2.0
kaitaistruct==0.8
Keras-Preprocessing==1.1.2
kiwisolver @ file:///C:/ci/kiwisolver_1604014703538/work
llvmlite==0.35.0
lxml==4.5.2
lz4==3.0.2
Mako==1.1.3
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib @ file:///C:/ci/matplotlib-base_1603355780617/work
mergedeep==1.3.0
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
mlflow==1.8.0
modes==1.0.2
mypy==0.790
mypy-extensions==0.4.3
netifaces==0.10.9
networkx @ file:///tmp/build/80754af9/networkx_1598376031484/work
numba==0.52.0
numpy @ file:///C:/ci/numpy_and_numpy_base_1603466732592/work
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
olefile==0.46
opt-einsum==3.3.0
opticalmaterialspy==0.3.3
OWSLib @ file:///home/conda/feedstock_root/build_artifacts/owslib_1607557132055/work
packaging==20.8
pandas==1.2.0
parsel==1.6.0
patsy==0.5.1
Pillow @ file:///C:/ci/pillow_1603823068645/work
pluggy==0.13.1
polaris-ml==0.9.4
prometheus-client==0.8.0
prometheus-flask-exporter==0.14.1
protobuf==3.14.0
ptflops==0.6.4
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1593275161868/work
pydub==0.24.1
pyepsg==0.4.0
pykdtree @ file:///D:/bld/pykdtree_1605125716861/work
pyOpenSSL @ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1608055815057/work
pyparsing==2.4.7
pypiwin32==223
pyproj==2.4.2.post1
pyrsistent==0.17.3
pyshp @ file:///home/conda/feedstock_root/build_artifacts/pyshp_1599782465740/work
PySocks @ file:///D:/bld/pysocks_1602327034683/work
pytest==6.2.1
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.1
PyWavelets @ file:///C:/ci/pywavelets_1601658407916/work
pywin32==228
PyYAML==5.3.1
querystring-parser==1.2.4
requests==2.24.0
requests-oauthlib==1.3.0
resampy==0.2.2
rsa==4.6
satnogs-decoders==1.7.0
scikit-image==0.17.2
scikit-learn==0.24.0
scipy==1.4.1
seaborn==0.11.1
selenium==3.141.0
Shapely==1.7.0
simplejson==3.17.0
six==1.15.0
smmap==3.0.4
SQLAlchemy==1.3.13
sqlparse==0.3.1
statsmodels==0.12.1
tabulate==0.8.7
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.4.0
tensorflow-estimator==2.4.0
termcolor==1.1.0
threadpoolctl==2.1.0
tifffile @ file:///home/conda/feedstock_root/build_artifacts/tifffile_1591280222285/work
toml==0.10.2
toolz @ file:///tmp/build/80754af9/toolz_1601054250827/work
torch==1.7.1+cu101
torchaudio==0.7.2
torchsummary==1.5.1
torchvision==0.8.2+cu101
tornado @ file:///C:/ci/tornado_1606942392901/work
tqdm==4.55.1
transforms3d==0.3.1
typed-ast==1.4.2
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1598376058250/work
urllib3 @ file:///tmp/build/80754af9/urllib3_1606938623459/work
w3lib==1.22.0
waitress==1.4.4
websocket-client==0.57.0
Werkzeug==1.0.1
win-inet-pton @ file:///D:/bld/win_inet_pton_1602260115377/work
wincertstore==0.2
wrapt==1.12.1
wurlitzer==2.0.1
xgboost==1.1.1
comet-ml==3.6.0
